Updated mean temperature code
Given the small size of these devices, they heat up a lot.The previous calculation can go so far as to predict many thousands of degrees for a several watt input power. This updated version accounts for the heat loss via radiation and produces results that are far more realistic. Furthermore, it is capable of estimating the natural-convection heat transfer coefficient aswell, incase the user has not imposed a value.
The code has also been checked for errors with the validation cases and is working well.