Modifications to mean temperature and efficiency functions (#3)
Given the small size of these devices, they heat up a lot. The previous calculation could go so far as to predict many thousands of degrees for a several watt input power. 

This updated version accounts for the heat loss via radiation and produces results that are far more realistic. Furthermore, it is capable of estimating the natural-convection heat transfer coefficient as well, in case the user has not imposed a value. The code has also been checked for errors with the validation cases and is working well.

* Updated the mean temperature code.

* Added the heat transfer coefficient to `ThermophoneSimOpts`.

* Removing `matlabFunction` in mean temperature function.

* Update `meanTemperature.m`.

* Update `efficiencyCalc.m`.

* Removed unused variables and unneeded comments and changed from using magnitude of the heat-flux to using the absolute values of the real component