Added the impact of non-dimensionalisation to gamma ratio
The result is now quite close, however, I am not convinced that this is the true gamma dependence - but this is the equation that Avshalom came up with in eq.43 http://dx.doi.org/10.1016/j.jsv.2015.08.016. It could also be that my estimation of thermal layer thickness is not that good, but this is not likely.